Version 2.3.7

+++ app/src/main/java/threads/server/provider/ProviderDataSource.java
@@ -41,7 +41,7 @@ public class ProviderDataSource extends BaseDataSource {
              byte[] data = fileReader.load(size);
              System.arraycopy(data, 0, buffer, offset, data.length);
 
-            return (int) data.length;
+            return data.length;
 
         } catch (Throwable e) {
             throw new IOException(e);
